Move multisig broker into its own folder
