+++ ironfish-cli/src/multisigBroker/README.md
@@ -1,0 +1,3 @@
+# Multisig Broker
+
+This is a server / client that allows multisig participants to broker creating DKG accounts and signing transactions through a trustless server.
